Rename log depth option and wire UI setting
Replaced `maxConsoleLogDepth`/`verbosity` with `maxLoggingDepth` across project and workflow execution options, and passed it through open/save project execution paths. Added a user-configurable "Maximum logging depth" setting to project open/save dialogs with persisted preferences. Also fixed minor numeric type issues by casting KDE values to float and using `std::fabs` in MeanShift bounds calculation.

+++ ManiVault/src/util/DensityComputation.cpp
@@ -296,7 +296,7 @@ float DensityComputation::calculateMaxKDE()
     float maxKDE = std::numeric_limits<float>::lowest();
     for (size_t i = 0; i < kde.size(); i += 3) // only lookup red channel
     {
-        maxKDE = std::max(maxKDE, kde[i]);
+        maxKDE = std::max(maxKDE, static_cast<float>(kde[i]));
     }
 
     return maxKDE;
